+++ src/Blitz.Avalonia.Controls/Views/ResultsBox.axaml.cs
@@ -27,7 +27,7 @@ public partial class ResultsBox : UserControl
         if (e.Key == Key.Enter)
         {
             var selected = mainWindowViewModel.SelectedItems.FirstOrDefault();
-            if (mainWindowViewModel.SelectedEditorViewModel != null && !mainWindowViewModel.SelectedEditorViewModel.RunTotoOnObjectGoto(selected,
+            if (mainWindowViewModel.SelectedEditorViewModel != null && !mainWindowViewModel.SelectedEditorViewModel.RunTotoOnObjectGoto(selected,true,
                     out string errorMessage))
             {
                 mainWindowViewModel.ShowImportantMessage?.Invoke(errorMessage);
@@ -43,7 +43,22 @@ public partial class ResultsBox : UserControl
             return;
         }
         
+        if (e.Source is Control control)
+        {
+            if (!mainWindowViewModel.SelectedEditorViewModel.RunTotoOnObjectGoto(e.AddedItems[0],true,
+                    out string errorMessage))
+            {
+                mainWindowViewModel.ShowImportantMessage?.Invoke(errorMessage);
+            }
+        }
+        
         mainWindowViewModel.ShowPreview?.Invoke(e.AddedItems[0]!);
+        
+        if (mainWindowViewModel.SelectedEditorViewModel == null)
+        {
+            return;
+        }
+
     }
     
     private  void GotoOtherEditorRun(int parameter)
@@ -51,7 +66,7 @@ public partial class ResultsBox : UserControl
         if (DataContext is not MainWindowViewModel mainWindowViewModel) return;
         var editorVm = mainWindowViewModel.GotoEditorCollection[parameter];
         var selectedFirst = mainWindowViewModel.SelectedItems.FirstOrDefault();
-        if (! editorVm.RunTotoOnObjectGoto(selectedFirst, out var errorMessage))
+        if (! editorVm.RunTotoOnObjectGoto(selectedFirst, false, out var errorMessage))
         {
              ShowImportantMessage(errorMessage);
         }
@@ -127,7 +142,7 @@ public partial class ResultsBox : UserControl
 
         if (e.Source is Control control)
         {
-            if (!mainWindowViewModel.SelectedEditorViewModel.RunTotoOnObjectGoto(control.DataContext,
+            if (!mainWindowViewModel.SelectedEditorViewModel.RunTotoOnObjectGoto(control.DataContext,false,
                     out string errorMessage))
             {
                 mainWindowViewModel.ShowImportantMessage?.Invoke(errorMessage);